requirements.txt
beautifulsoup4==4.8.1
requests==2.22.0
tqdm==4.40.0
aiohttp==3.6.2
python-dateutil==2.8.1